Rahu 313 (#27)
* IFIX-311: Create Department Hierarchy

* Merge branch 'ifix-department-entity-service' of https://github.com/egovernments/iFix-Dev into pk-ifix-department-entity-service

# Conflicts:
#	domain-services/ifix-department-entity-service/src/main/resources/application.properties

* IFIX-313: Department entity create api

* changed the level data tpe to Integer

* DepartmentEntity attribute correction

* refactor the GovernmentUtil and DepartmentUtil

* DepartmentEntity validation fix

+++ domain-services/ifix-department-entity-service/src/main/java/org/egov/web/models/DepartmentHierarchyLevel.java
@@ -37,7 +37,7 @@ public class DepartmentHierarchyLevel {
     private String parent = null;
 
     @JsonProperty("level")
-    private BigDecimal level = null;
+    private Integer level = null;
 
     @JsonProperty("auditDetails")
     private AuditDetails auditDetails = null;
